// frontend/src/api/models/DynamicPaginatedResponse.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ItemSchema } from '../models/ItemSchema';
import { HttpFile } from '../http/http';

export class DynamicPaginatedResponse {
    'data': Array<ItemSchema>;
    'totalCount': number;
    'hasMore': boolean;
    'page': number | null;
    'itemsPerPage': number | null;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<ItemSchema>",
            "format": ""
        },
        {
            "name": "totalCount",
            "baseName": "total_count",
            "type": "number",
            "format": ""
        },
        {
            "name": "hasMore",
            "baseName": "has_more",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "page",
            "baseName": "page",
            "type": "number",
            "format": ""
        },
        {
            "name": "itemsPerPage",
            "baseName": "items_per_page",
            "type": "number",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return DynamicPaginatedResponse.attributeTypeMap;
    }

    public constructor() {
    }
}