// frontend/src/api/models/DynamicListResponse.ts
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { ItemSchema } from '../models/ItemSchema';
import { HttpFile } from '../http/http';

export class DynamicListResponse {
    'data': Array<ItemSchema>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "data",
            "baseName": "data",
            "type": "Array<ItemSchema>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return DynamicListResponse.attributeTypeMap;
    }

    public constructor() {
    }
}